final_project/backend/requirements.txt
flask==2.2.3
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
moviepy==1.0.3
SpeechRecognition==3.10.0
pymongo==4.3.3
nltk==3.9.1
scikit-learn==1.5.2
google-api-python-client==2.86.0
python-dotenv==1.0.0
pyjwt==2.6.0
pytest==7.3.1
twilio==8.2.0
schedule==1.2.0
